LayoutTests/plugins/plugin-javascript-access-expected.txt

Name: WebKit Test PlugIn

Description: Simple Netscape plug-in that handles test content for WebKit

Filename: TestNetscapePlugIn.plugin

Mime Types:

Type: application/x-webkit-test-netscape

Description: test netscape content

Suffixes: testnetscape


PluginArray.item() works.

PluginArray.namedItem() works.

Type: application/x-webkit-test-netscape

Description: test netscape content

Suffixes: testnetscape

MimeTypeArray.item() works.

MimeTypeArray.namedItem() works.
